Seat reservation

+++ backend/DTOs/LoginResponseDTO.cs
@@ -6,6 +6,6 @@ namespace PublicTransportNavigator.DTOs
     {
         public required UserDTO User { get; set; }
         public required string Token { get; set; }
-        public required int ExpirationTime { get; set; }
+        public required long ExpirationTime { get; set; }
     }
 }
