+++ backend/DTOs/RouteDetailsDTO.cs
@@ -7,7 +7,7 @@ namespace PublicTransportNavigator.DTOs
         public string Id { get; set; }
         public TimeSpan DepartureTime { get; set; }
         public TimeSpan DestinationTime { get; set; }
-        public int TravelTime { get; set; }
+        public long TravelTime { get; set; }
         public List<RoutePartDTO> Parts { get; set; } = [];
         public List<Coordinate> Coordinates { get; set; } = [];
     }
